started reworking the interface, iroModule class
