+++ src/iroModule.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <iro.hpp>
+#include <util/module.hpp>
+
+class iroModule : public module
+{
+public:
+
+};
